requirements.txt
discord.py>=2.3.0
pyyaml>=6.0
libtmux>=0.28.0
